pyproject.toml
[project]
name = "openapi-to-mcp"
version = "0.1.0"
description = "Convert OpenAPI/Swagger specs to MCP servers - expose any REST API as Claude tools"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Job Nijenhuis" }]
keywords = ["mcp", "openapi", "swagger", "claude", "anthropic", "api", "llm", "tools"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/afstkla/openapi-to-mcp"
Repository = "https://github.com/afstkla/openapi-to-mcp"
Issues = "https://github.com/afstkla/openapi-to-mcp/issues"

[project.scripts]
openapi-to-mcp = "openapi_to_mcp.cli:main"

[dependency-groups]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-httpx>=0.32.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openapi_to_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
